refactor: 하이픈 자동 완성 매서드 분리
